backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
pydantic-settings
deepface
opencv-python
numpy
python-socketio[asgi]
